Don't try to remove non-existent files (oops)

+++ utils.py
@@ -1406,7 +1406,8 @@ def ABCToPostscript(abc, svg=False, eps=False, target=None):
         if target is None:
             target = os.path.splitext(fn)[0] + '.ps'
             
-    os.unlink(target)
+    if os.path.exists(target):
+        os.unlink(target)
     
     cmdline = [kExecutable, fn, svg_arg, '-O', target]
     cmdline = ' '.join(cmdline)
